Leaflet Engine
Added the 'engine' library. It allow to do the interface between the
editor and leaflet.
It allow to create a map and several layers to add to the map.
Layer in this library correspond to a sort of leaflet LayerGroup.
So, you can create a Layer, add it to the map, and then, you can add to
this Layer, several leaflet objetcs (marker, polygon, geoJSON Layers and
so on).

It is just designed to gather several layers between themselves and have
others properties like a name and a description for this group of
leaflet layers.